+++ NEWS.md
@@ -1,3 +1,5 @@
 # dsn 0.0.1
 
+* New functions `sds()` and `vrtcon()`. 
+
 * First basic functions.